requirements.txt
torch
torchvision
pandas
numpy
opencv-python
scikit-learn
pillow
streamlit